Backend/thebinaries/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-grappelli==2.15.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
# pkg_resources==0.0.0
